Create an abstract class for inheriting sorting logic to BatchCollect
